Refactor optimization to a separate domain
